requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client==2.2.4
langchain==0.3.19
flask==3.1.0
pypdf==5.3.0
python-dotenv==1.0.1
-e .